% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_deri.R
\name{catvar_clust}
\alias{catvar_clust}
\title{doing discrete variables clustering}
\usage{
catvar_clust(ldf, m)
}
\arguments{
\item{ldf}{the dataframe}

\item{m}{the feature numbers}

\item{k}{cluster numbers}
}
\description{
doing discrete variables clustering
}
\examples{

catvar_clust(ldf,m)
}

