% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_deri.R
\name{df_cluster}
\alias{df_cluster}
\title{doing discrete variables clustering}
\usage{
df_cluster(ldf, k)
}
\arguments{
\item{ldf}{the dataframe}

\item{k}{cluster numbers}

\item{m}{the feature numbers}
}
\description{
doing discrete variables clustering
}
\examples{

discrete_vars_cluster(ldf,k,m)
}

