% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_deri.R
\name{find_cluster_bstpara}
\alias{find_cluster_bstpara}
\title{select best cluster paras}
\usage{
find_cluster_bstpara(ldf, kk, target, m = NULL)
}
\arguments{
\item{ldf}{the dataframe}

\item{kk}{cluster numbers}

\item{target}{the target variable}

\item{mm}{the feature numbers}
}
\description{
select best cluster paras
}
\examples{

find_cluster_bstpara(ldf,kk,mm,target)
}

