% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nums_handle.R
\name{get_inc}
\alias{get_inc}
\title{get incline of a vector}
\usage{
get_inc(value, type = "n1")
}
\arguments{
\item{value}{the discreting dataframe by best seperation}

\item{type}{the vars need to change}
}
\description{
This function will get incline of a vector
}
\examples{
get_inc(value,type='n1')
}

