% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nums_handle.R
\name{get_num_statis}
\alias{get_num_statis}
\title{get numeric variables' statistics index}
\usage{
get_num_statis(df, key.vec, varname, targ.vec = NULL)
}
\arguments{
\item{df}{the discreting dataframe by best seperation}

\item{key.vec}{the vars need to change}

\item{varname}{the label variable}

\item{reordervar}{whether reorder the dataframe by media}
}
\description{
This function will get numeric variables' statistics index
}
\examples{
get_num_statis(alldata,alldata[,'Idx'],'varname',reordervar=F)
}

