% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get rules info.R
\name{get_rule_hit}
\alias{get_rule_hit}
\title{get a rule hits variables}
\usage{
get_rule_hit(df, rule.list, min.odds, min.num)
}
\arguments{
\item{df}{the dataframe}

\item{rule.list}{the rule list}
}
\description{
get a rule hits variables
}
\examples{
get_rule_hit(df,rule.list)
}

