% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get rules info.R
\name{get_rules}
\alias{get_rules}
\title{get the special rules}
\usage{
get_rules(df, targ)
}
\arguments{
\item{df}{the dataframe}

\item{targ}{the target variable}

\item{min.odds}{the minimum bad odds level}

\item{min.num}{the minimum bad obs numbers}
}
\description{
get the special rules
}
\examples{

get_rules(df,targ,min.odds,min.num)
}

