% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get the stable difference matrix.R
\name{get_sd_probs}
\alias{get_sd_probs}
\title{transform the stable difference matrix}
\usage{
get_sd_probs(rowj, df, y, njs, rowis = NULL)
}
\arguments{
\item{out.list}{the Markov lineage}

\item{kk}{the target vect}

\item{targ.vec}{lowest nums}
}
\description{
transform the stable difference matrix
}
\examples{

get_sd_vec(out.list,kk,targ.vec)
}

