% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get the stable difference matrix.R
\name{get_std_deri}
\alias{get_std_deri}
\title{fit the stable difference matrix}
\usage{
get_std_deri(df, loop = 2, nsmp = 3000, threshold = 0.2)
}
\arguments{
\item{train.mt}{the Markov lineage}

\item{pred.mt}{the target vect}

\item{targ.vec}{lowest nums}
}
\description{
fit the stable difference matrix
}
\examples{

stable_difference(train.mt,pred.mt,targ.vec)
}

