% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_func.R
\name{get_the_mk_vec}
\alias{get_the_mk_vec}
\title{transform the markov fit}
\usage{
get_the_mk_vec(markov_mat, ddr, train.y, test.y, lowest.vec, diff.low)
}
\arguments{
\item{ddr}{whether delete duplicate rows}

\item{train.y}{the train target}

\item{test.y}{the test target}

\item{lowest.vec}{best lowest}

\item{gb}{the Markov lineage matrix}
}
\description{
transform the markov fit
}
\examples{

get_the_mk_vec(gb,ddr=F,targ.vec,target.y,1:300)
}

