% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping_deri.R
\name{group_base}
\alias{group_base}
\title{get the base aggregation values}
\usage{
group_base(idx.vec, value.vec, var.name)
}
\arguments{
\item{idx.vec}{s}

\item{value.vec}{s}

\item{var.name}{s}
}
\description{
get the base aggregation values
}
\examples{

group_base(idx.vec,value.vec,'log1')
}

