% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping_deri.R
\name{group_cat_time}
\alias{group_cat_time}
\title{get the advanced aggregation values with time and status}
\usage{
group_cat_time(idx.vec, value.vec, time.vec, targ.vec)
}
\arguments{
\item{idx.vec}{s}

\item{value.vec}{s}

\item{time.vec}{s}

\item{targ.vec}{s}
}
\description{
get the advanced aggregation values with time and status
}
\examples{

group_cat_time(idx.vec,value.vec,time.vec,targ.vec)
}

