% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping_deri.R
\name{group_time}
\alias{group_time}
\title{get the advanced aggregation values with time}
\usage{
group_time(idx.vec, time.vec, targ.vec)
}
\arguments{
\item{idx.vec}{s}

\item{time.vec}{s}

\item{targ.vec}{s}
}
\description{
get the advanced aggregation values with time
}
\examples{

group_time(idx.vec,time.ve,targ.vec)
}

