% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_func.R
\name{markov_fit}
\alias{markov_fit}
\title{fit the Markov}
\usage{
markov_fit(markov.mat, targ.vec, lowest, ddr = T)
}
\arguments{
\item{markov.mat}{the Markov lineage}

\item{targ.vec}{the target vect}

\item{lowest}{lowest nums}

\item{ddr=F}{ddr}
}
\description{
fit the Markov
}
\examples{

markov_fit(markov.mat,targ.vec,30,ddr=T)
}

