% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_func.R
\name{markov_transform}
\alias{markov_transform}
\title{transform the markov fit}
\usage{
markov_transform(markov.mat, markov.fit, targ.vec)
}
\arguments{
\item{markov.mat}{the Markov lineage}

\item{markov.fit}{the fitted markov}

\item{targ.vec}{the target vect}

\item{init}{init}
}
\description{
transform the markov fit
}
\examples{

markov_transform(markov.mat,markov.fit,targ.vec,init=F)
}

