% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get rules info.R
\name{match_oris}
\alias{match_oris}
\title{find those vars which match the bisic var structures}
\usage{
match_oris(oris, vars, l)
}
\arguments{
\item{oris}{basic var structure}

\item{vars}{complate vars}

\item{l}{the place which don't care}
}
\description{
find those vars which match the bisic var structures
}
\examples{

match_orisv(oris,vars,l)
}

