% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cate_handle.R
\name{one_hot_transform}
\alias{one_hot_transform}
\title{change discrete variable to bad odds}
\usage{
one_hot_transform(df, targ, min.rat = 0.05)
}
\arguments{
\item{df}{the dataframe}

\item{targ}{the label variable}

\item{min.rat}{the minimal sample}
}
\description{
This function will change discrete variable to bad odds
}
\examples{
one_hot_transform(alldata,'target',0.1)
}

