% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cate_handle.R
\name{value2interchanged_num}
\alias{value2interchanged_num}
\title{interchanged switch average y}
\usage{
value2interchanged_num(df, targ, fold, avgfunc)
}
\arguments{
\item{df}{the discreting dataframe by best seperation}

\item{targ}{the label variable}

\item{fold}{the subsample numbers}
}
\description{
This function will interchanged switch average y
}
\examples{
value2interchanged_num(alldata,'target',4,median)
}

