% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cate_handle.R
\name{value2interchanged_odds}
\alias{value2interchanged_odds}
\title{interchanged switch the odds in different subsamples}
\usage{
value2interchanged_odds(df, targ, fold)
}
\arguments{
\item{df}{the discreting dataframe by best seperation}

\item{targ}{the label variable}

\item{fold}{the subsample numbers}
}
\description{
This function will interchanged switch the odds in different subsamples
}
\examples{
value2interchanged_odds(alldata,'target',4)
}

