% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cate_handle.R
\name{value2odds}
\alias{value2odds}
\title{change discrete variable to bad odds}
\usage{
value2odds(df, vars, targ, lowest = 50)
}
\arguments{
\item{df}{the discreting dataframe by best seperation}

\item{vars}{the vars need to change}

\item{targ}{the label variable}
}
\description{
This function will change discrete variable to bad odds
}
\examples{
value2odds(df,vars,targ,lowest=50)
}

